Move publishing to org.korge and update Kotlin dependencies to 2.3.20
- Toolchain was updated: Kotlin 2.3.20, AGP 8.13.1, Dokka 2.1.0, Gradle 8.14.4
- Custom Sonatype staging/release logic was replaced with
  com.vanniktech.maven.publish for signing + Central Portal publishing
- Namespace on maven central was changed to org.korge and Maven Central
  badge in README.md was updated
- GitHub Actions workflows were refreshed (Java 21, updated actions versions,
  new publish and dokka tasks, optional yarn lock upgrade)
- API check was updated
- Update korlibs dependencies:
    org.korge:korlibs-checksum:6.1.0
    org.korge:korlibs-compression:6.1.0
    org.korge:korlibs-jseval:6.1.0
    org.korge:korlibs-io-vfs:6.1.0
    org.korge:korlibs-io-stream:6.1.0
    org.korge:korlibs-io-network-core:6.1.0
    org.korge:korlibs-math-core:6.1.0
    org.korge:korlibs-memory:6.1.0
    org.korge:korlibs-ffi-legacy:6.1.1
    org.korge:korlibs-crypto:6.1.0
    org.korge:korlibs-encoding:6.1.0
    org.korge:korlibs-platform:6.1.0
    org.korge:korlibs-datastructure:6.1.0
    org.korge:korlibs-number:6.1.0
    org.korge:korlibs-time:6.1.0
    org.korge:korlibs-logger:6.1.0
    org.korge:korlibs-dyn:6.1.0
    org.korge:korlibs-string:6.1.0
    org.korge:korlibs-serialization:6.1.0
    org.korge:korlibs-io-fs:6.1.0
    org.jetbrains.kotlinx:atomicfu:0.32.1
    org.jetbrains.kotlinx:kotlinx-coroutines-test:1.10.2

+++ korlibs-io/src/korlibs/io/net/URL.kt
@@ -210,7 +210,7 @@ data class URL private constructor(
 			//for (byte c : data) System.out.printf("%02X\n", c & 0xFF);
 			for (n in 0 until data.size) {
 				val c = data[n]
-				val cc = c.toChar()
+				val cc = c.toInt().toChar()
 				when (cc) {
 					' ' -> if (formUrlEncoded) sb.append("+") else sb.append("%20")
 					in 'a'..'z', in 'A'..'Z', in '0'..'9', '-', '_', '.', '*' -> sb.append(cc)
